+++ orm/model/ormtable/auto_increment_test.go
@@ -52,7 +52,7 @@ func runAutoIncrementScenario(t *testing.T, table ormtable.AutoIncrementTable, c
 	ex1 := &testpb.ExampleAutoIncrementTable{X: "foo", Y: 5}
 	assert.NilError(t, store.Save(ctx, ex1))
 	assert.Equal(t, uint64(1), ex1.Id)
-	curSeq, err := table.CurrentSequence(ctx)
+	curSeq, err := table.LastInsertedSequence(ctx)
 	assert.NilError(t, err)
 	assert.Equal(t, curSeq, uint64(1))
 
@@ -61,7 +61,7 @@ func runAutoIncrementScenario(t *testing.T, table ormtable.AutoIncrementTable, c
 	assert.NilError(t, err)
 	assert.Equal(t, uint64(2), ex2.Id)
 	assert.Equal(t, newId, ex2.Id)
-	curSeq, err = table.CurrentSequence(ctx)
+	curSeq, err = table.LastInsertedSequence(ctx)
 	assert.NilError(t, err)
 	assert.Equal(t, curSeq, uint64(2))
 
@@ -84,7 +84,7 @@ func runAutoIncrementScenario(t *testing.T, table ormtable.AutoIncrementTable, c
 	ex1.Id = 0
 	assert.NilError(t, table.Insert(store3, ex1))
 	assert.Equal(t, uint64(3), ex1.Id) // should equal 3 because the sequence number 2 should have been imported from JSON
-	curSeq, err = table.CurrentSequence(store3)
+	curSeq, err = table.LastInsertedSequence(store3)
 	assert.NilError(t, err)
 	assert.Equal(t, curSeq, uint64(3))
 }